got the enum export working and separating out types that will traverse modules
