+++ apps/mmessier/app/projects/use-query-parameters-demo/[routeName]/style.ts
@@ -1,6 +1,7 @@
 import { colors } from '../../../styles/colors';
-import { Breakpoint, Style } from '../../../../types';
+import { Breakpoint, Style } from '@mmessier/types';
 import { spacingLevel } from '../../../styles/spacing';
+
 export const styles: Style = (breakpoint: Breakpoint) => ({
   static: {
     container: {
