+++ package.json
@@ -47,6 +47,8 @@
     "@nx/workspace": "17.1.3",
     "@rollup/plugin-url": "^7.0.0",
     "@svgr/rollup": "^8.0.1",
+    "@swc-node/register": "~1.6.7",
+    "@swc/core": "~1.3.85",
     "@testing-library/react": "14.0.0",
     "@types/jest": "^29.4.0",
     "@types/lodash": "^4.14.199",
