+++ types/README.md
@@ -1,0 +1,7 @@
+# types
+
+This library was generated with [Nx](https://nx.dev).
+
+## Building
+
+Run `nx build types` to build the library.
