+++ types/project.json
@@ -1,0 +1,26 @@
+{
+  "name": "types",
+  "$schema": "../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "types/src",
+  "projectType": "library",
+  "targets": {
+    "build": {
+      "executor": "@nx/js:tsc",
+      "outputs": ["{options.outputPath}"],
+      "options": {
+        "outputPath": "dist/types",
+        "main": "types/src/index.ts",
+        "tsConfig": "types/tsconfig.lib.json",
+        "assets": ["types/*.md"]
+      }
+    },
+    "lint": {
+      "executor": "@nx/eslint:lint",
+      "outputs": ["{options.outputFile}"],
+      "options": {
+        "lintFilePatterns": ["types/**/*.ts", "types/package.json"]
+      }
+    }
+  },
+  "tags": []
+}
