+++ types/src/lib/styles.ts
@@ -1,0 +1,28 @@
+import { SxProps, Theme } from '@mui/material';
+import { CSSProperties } from 'react';
+
+export const enum Breakpoint {
+  SMALL = 'small',
+  MEDIUM = 'medium',
+  DEFAULT = 'default',
+}
+
+type StaticStyle = {
+  [x: string]: CSSProperties;
+};
+
+type DynamicStyle = {
+  [y: string]: (...args) => CSSProperties;
+};
+
+type SXStyle = {
+  [z: string]: SxProps<Theme>;
+};
+
+export type BasicStyle = {
+  dynamic?: DynamicStyle;
+  static?: StaticStyle;
+  sx?: SXStyle;
+};
+
+export type Style = (breakpoint: Breakpoint) => BasicStyle;
